+++ include/network/ipv4address.h
@@ -1,3 +1,25 @@
+/**
+ * @file
+ * @author Nicholas Logozzo <nlogozzo225@gmail.com>
+ *
+ * @section LICENSE
+ *
+ * This program is free software; you can redistribute it and/or
+ * modify it under the terms of the GNU General Public License as
+ * published by the Free Software Foundation; either version 3 of
+ * the License, or (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful, but
+ * WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU
+ * General Public License for more details at
+ * https://www.gnu.org/copyleft/gpl.html
+ *
+ * @section DESCRIPTION
+ *
+ * A model of an IPv4 address.
+ */
+
 #ifndef IPV4ADDRESS_H
 #define IPV4ADDRESS_H
 
@@ -20,6 +42,11 @@ namespace Nickvision::Network
          * @param fourth The fourth byte of the address
          */
         IPv4Address(unsigned char first, unsigned char second, unsigned char third, unsigned char fourth);
+        /**
+         * @brief Constructs an IPv4Address.
+         * @param address The address as a long
+         */
+        IPv4Address(unsigned long address);
         /**
          * @brief Gets the first byte of the address.
          * @return The first byte of the address